// src/shared/api/generated/types/FieldFilter.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Фильтр по полю
 * @example [object Object]
 */
export type FieldFilter = {
  /**
   * @description Имя поля
   * @type string
   */
  field_name: string
  /**
   * @description Оператор сравнения
   * @type string
   */
  operator: string
  /**
   * @description Значения для сравнения
   * @type array
   */
  values: (boolean | number | number | string)[]
  /**
   * @description Активен ли фильтр
   * @default true
   * @type boolean | undefined
   */
  is_active?: boolean
}